+++ frontends/web/src/routes/account/send/send.tsx
@@ -29,7 +29,7 @@ import { HideAmountsButton } from '@/components/hideamountsbutton/hideamountsbut
 import { Button } from '@/components/forms';
 import { BackButton } from '@/components/backbutton/backbutton';
 import { Column, ColumnButtons, Grid, GuideWrapper, GuidedContent, Header, Main } from '@/components/layout';
-import { Amount } from '@/components/amount/amount';
+import { AmountWithUnit } from '@/components/amount/amount-with-unit';
 import { FeeTargets } from './feetargets';
 import { isBitcoinBased } from '@/routes/account/utils';
 import { ConfirmSend } from './components/confirm/confirm';
@@ -43,6 +43,7 @@ import { FiatValue } from './components/fiat-value';
 import { TProposalError, txProposalErrorHandling } from './services';
 import { CoinControl } from './coin-control';
 import { connectKeystore } from '@/api/keystores';
+import { SubTitle } from '@/components/title';
 import style from './send.module.css';
 
 type TProps = {
@@ -380,19 +381,18 @@ export const Send = ({
           </Header>
           <View>
             <ViewContent>
-              <div>
-                <label className="labelXLarge">{t('send.availableBalance')}</label>
-              </div>
-              <Balance balance={balance} noRotateFiat/>
-              <div className={`flex flex-row flex-between ${style.container || ''}`}>
-                <label className="labelXLarge">{t('send.transactionDetails')}</label>
-                <div className={style.coinControlButtonContainer}>
-                  <CoinControl
-                    account={account}
-                    onSelectedUTXOsChange={handleSelectedUTXOsChange}
-                    onCoinControlDialogActiveChange={setUtxoDialogActive}
-                  />
+              <div className={style.sendHeader}>
+                <div className={style.availableBalance}>
+                  <Balance balance={balance} noRotateFiat/>
                 </div>
+                <SubTitle className={style.subTitle}>
+                  {t('send.transactionDetails')}
+                </SubTitle>
+                <CoinControl
+                  account={account}
+                  onSelectedUTXOsChange={handleSelectedUTXOsChange}
+                  onCoinControlDialogActiveChange={setUtxoDialogActive}
+                />
               </div>
               <Grid col="1">
                 <Column>
@@ -494,16 +494,21 @@ export const Send = ({
                 onContinue={handleContinue}
                 onRetry={handleRetry}>
                 <p>
-                  {(proposedAmount &&
-                  <Amount alwaysShowAmounts amount={proposedAmount.amount} unit={proposedAmount.unit}/>) || 'N/A'}
-                  {' '}
-                  <span className={style.unit}>
-                    {(proposedAmount && proposedAmount.unit) || 'N/A'}
-                  </span>
+                  {proposedAmount && (
+                    <AmountWithUnit
+                      amount={proposedAmount}
+                      alwaysShowAmounts
+                      unitClassName={style.unit}
+                    />
+                  )}
+                  <br />
+                  {(proposedAmount && proposedAmount.conversions && proposedAmount.conversions[activeCurrency]) ? (
+                    <FiatValue
+                      amount={proposedAmount.conversions[activeCurrency] || ''}
+                      baseCurrencyUnit={activeCurrency}
+                    />
+                  ) : null}
                 </p>
-                {(proposedAmount && proposedAmount.conversions && proposedAmount.conversions[activeCurrency]) ? (
-                  <FiatValue baseCurrencyUnit={activeCurrency} amount={proposedAmount.conversions[activeCurrency] || ''} />
-                ) : null}
               </SendResult>
             )}
           </View>